Extract Mutation module
