+++ hydra-node/hydra-node.cabal
@@ -228,6 +228,7 @@ test-suite tests
     Hydra.API.ServerSpec
     Hydra.APISpec
     Hydra.BehaviorSpec
+    Hydra.Chain.Direct.Contract.Mutation
     Hydra.Chain.Direct.ContractSpec
     Hydra.Chain.Direct.Fixture
     Hydra.Chain.Direct.MockServer
